refactor(qbittorrent-e2e): extract login readiness step
